% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgpr.R
\name{mgpr}
\alias{mgpr}
\title{Fitting an mgpr model}
\usage{
mgpr(
  datay,
  datax,
  kernel = "matern32",
  kernpar = list(sigma = NULL, corlen = NULL, errorvar = NULL),
  meanf = "avg",
  optimpar = list(optkfold = 5, optstart = c(1, 10, 0.1), optlower = c(0.3, 3, 0.03),
    optupper = c(10, 50, 0.5), optcontrol = list(t0 = 10, nlimit = 50, r = 0.9)),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{datay}{a data frame or an object coercible by as.data.frame to a
data frame containing the response variable or variables (columns) for each
observation (rows). It is recommended to use row and column names.}

\item{datax}{a data frame or an object coercible by as.data.frame to a
data frame containing the predictor variables (columns) for each observation
(rows). It is recommended to use row and column names.}

\item{kernel}{a name of the kernel function.
Alternatives: "\code{matern12}", "\code{matern32}" (default), 
"\code{matern52}", and
"\code{rbf}".}

\item{kernpar}{kernel hyperparameters \code{sigma}, correlation 
length \code{corlen} and noise variance
\code{errorvar} given as a list.
Kernel hyperparameters not set by the user are 
estimated using the training set.}

\item{meanf}{mean function, choices: "\code{zero}", "\code{avg}" (default),
 "\code{linear}".
Mean function specifies a deterministic trend in the Gaussian process and mostly
affects extrapolation outside the training set. "\code{zero}" (default)
assumes no trend, "\code{avg}" uses the training set mean, "\code{linear}"
fits a linear minimum mean square error model to the training set.}

\item{optimpar}{List of control parameters for the hyperparameter estimation:
\code{optkfold} defines a kfold value for the cost function 
(must be in the range 2 and nrow(data)), \code{optstart} defines an initial
 vector for the optimization,
\code{optlower} defines a lower bound vector, \code{optupper} defines 
a upper bound vector, 
and \code{optcontrol} includes the control parameters of the simulated 
annealing used 
as an optimization approach in the hyperparameter estimation.
Initial and bound vectors have the order \code{c(sigma, corlen, errorvar)}.
For the control parameters, please refer to the documentation of the 
\code{optimization::optim_sa} function. 
Empty list uses the default parameters
defined in the \code{optimization::optim_sa} function (ver. 1.0-9). 
The default setup of
\code{mgpr} modifies three of the \code{optimization::optim_sa} function's 
default parameters: \code{optcontrol = list(t0 = 10, nlimit = 50, r = 0.9)}.}

\item{verbose}{logical. Report extra information on progress.}

\item{...}{additional arguments.}
}
\value{
\code{mgpr} returns an object of class "\code{mgpr}". The function
\code{\link{summary}} prints a summary of an \code{mgpr} object and the
function \code{\link{predict}} is used to make predictions with it.

An object of class "\code{mgpr}" is a list containing at least the following
components: \tabular{llllllllllllll}{
  \code{trainy} \tab \tab response variables of training data. \cr
  \code{trainx} \tab \tab predictor variables of training data. \cr
  \code{trainMeanSd} \tab \tab mean and standard deviation of 
  training data. \cr
  \code{Cy} \tab \tab sample covariance for response variables. \cr
  \code{E} \tab \tab error matrix. \cr
  \code{Ie} \tab \tab identity matrix for setting E for all observations. \cr
  \code{K} \tab \tab kernel matrix. \cr
  \code{kernel} \tab \tab a name of the kernel function. \cr
  \code{sigma} \tab \tab value of sigma. \cr
  \code{corlen} \tab \tab value of correlation length. \cr
  \code{errorvar} \tab \tab errorvar noise variance 
  (a.k.a. nugget parameter). \cr
  \code{meanf} \tab \tab name of the mean function \cr
  \code{XColName} \tab \tab column names of predictor variables. \cr
  \code{YColName} \tab \tab column names of response variable(s). \cr
  \code{XRowName} \tab \tab row names of predictor variables. \cr
}
}
\description{
\code{mgpr} is used to fit a Multivariate Gaussian Process Regression model.
}
\examples{
data(mgprdata)
m <- mgpr(
  datay = mgprdata[, 1:3], datax = mgprdata[, 4:39],
  kernel = "matern32", kernpar = list(sigma = 1, corlen = 5, errorvar = 0.1)
)
}
\references{
Varvia, P., Lähivaara, T., Maltamo, M., Packalen, P. 
and Seppänen, A. (2019). Gaussian process regression for forest attribute 
estimation from airborne laser scanning data. 
IEEE Transactions on Geoscience and Remote Sensing 57(6): 3361-3369.
https://doi.org/10.1109/TGRS.2018.2883495

Varvia, P., Räty, J. and Packalen, P. (2023). mgpr: An R package for
multivariate Gaussian process regression, SoftwareX 24: 101563.
https://doi.org/10.1016/j.softx.2023.101563
}
