% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPR110322.R
\name{predict}
\alias{predict}
\title{Predict method for MGPR model}
\usage{
predict(
  mgpr,
  newdatax = NULL,
  credinter = NULL,
  covout = FALSE,
  kfold = 0,
  fixneg = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{mgpr}{a \code{mgpr} model object, usually, a result of a call to
\code{\link{mgpr}}.}

\item{newdatax}{a data frame or an object coercible by as.data.frame to a
data frame containing new predictor variables. Column names must match with
the ones in the \code{mgpr} object. If NULL \code{datax} used in model
training is used.}

\item{credinter}{threshold of the credible intervals, e.g. \code{credinter=0.95}.
If NULL credible intervals are not returned.}

\item{covout}{logical. Determines whether the covariance matrix is returned
or not.}

\item{kfold}{value >=2 enables the k-fold cross validation mode. The
\code{kfold} argument defines the k value. The value must be in the range
2 and nrow(data).}

\item{fixneg}{logical. Determines whether negative predictions (and credible
intervals) are corrected to be positive or zero, see details in Varvia et al.
(2019).}

\item{verbose}{logical. Report extra information on progress.}
}
\value{
Returns predicted values. If credible intervals/covariance matrix
is/are requested, the output will be a list.
}
\description{
Predict using a Multivariate Gaussian process regression (MGPR) model. The
function supports k-fold cross validation, new predictor variables, limiting
predictions to positive scale, and generation of credible intervals.
}
\examples{
data(mgprdata)
m <- mgpr(
  datay = mgprdata[, 1:5], datax = mgprdata[, 6:29],
  kernel = "matern32", kernpar = list(sigma = 1, corlen = 5, errorvar = 0.1)
)
p <- predict(m, credinter = 0.95)
str(p)
}
\references{
Varvia, P., Lahivaara, T., Maltamo, M., Packalen, P. and Seppanen,
A. 2019. Gaussian process regression for forest attribute estimation from
airborne laser scanning data. 
IEEE Transactions on Geoscience and Remote Sensing 57(6): 3361-3369.

R package article to appear.
}
