% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPR110322.R
\name{summary}
\alias{summary}
\title{Summarizing MGPR model}
\usage{
summary(mgpr)
}
\arguments{
\item{mgpr}{a \code{mgpr} model object, usually, a result of a call to \code{\link{mgpr}}.}
}
\value{
A summary of the MGPR model will be printed to the R console.
}
\description{
\code{summary} method for class "\code{mgpr}".
}
\examples{
data(mgprdata)
m <- mgpr(
  datay = mgprdata[, 1:5], datax = mgprdata[, 6:29],
  kernel = "matern32", kernpar = list(sigma = 1, corlen = 5, errorvar = 0.1)
)
summary(m)
}
