% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MGPR170222.R
\name{mgpr}
\alias{mgpr}
\title{Fitting a MGPR model}
\usage{
mgpr(
  datay,
  datax,
  kernel,
  kernpar = "auto",
  meanf = "zero",
  optimpar = list(optkfold = 5, optstart = c(1, 10, 0.1), optlower = c(0.3, 3, 0.03),
    optupper = c(10, 50, 0.5)),
  verbose = FALSE,
  ...
)
}
\arguments{
\item{datay}{a data frame or an object coercible by as.data.frame to a 
data frame containing the response variable or variables (columns) for each 
observation (rows). It is recommended to use row and column names.}

\item{datax}{a data frame or an object coercible by as.data.frame to a
data frame containing the predictor variables (columns) for each observation
(rows). It is recommended to use row and column names.}

\item{kernel}{a name of the kernel function.
Alternatives: "\code{matern12}", "\code{matern32}", "\code{matern52}", and 
"\code{rbf}". A user must always define the kernel function.}

\item{kernpar}{kernel hyperparameters sigma, correlation length and noise variance 
(nugget) given as a list.
If kernpar is not given, kernel hyperparameters are estimated using the training set.}

\item{meanf}{mean function, choices: "\code{zero}", "\code{avg}", "\code{linear}".
Mean function specifies a deterministic trend in the Gaussian process and mostly 
affects extrapolation outside the training set. "\code{zero}" (default) 
assumes no trend, "\code{avg}" uses the training set mean, "\code{linear}" 
fits a linear minimum mean square error model to the training set.}

\item{optimpar}{List of control parameters for the hyperparameter estimation:
\code{optkfold} kfold value for the cost function, must be in the range 
2 and nrow(data), \code{optstart} initial vector for the optimization, 
\code{optlower} lower bound vector and \code{optupper} upper bound vector.
Initial and bound vectors have the order \code{c(sigma, corlen, errorvar)}.}

\item{verbose}{logical. Report extra information on progress.}

\item{...}{additional arguments.}
}
\value{
\code{mgpr} returns an object of class "\code{mgpr}". The function
\code{\link{summary}} prints a summary of an \code{mgpr} object and the 
function \code{\link{predict}} is used to make predictions with it.

An object of class "\code{mgpr}" is a list containing at least the following
components: \tabular{llllllllllllll}{
  \code{trainy} \tab \tab response variables of training data. \cr
  \code{trainx} \tab \tab predictor variables of training data. \cr
  \code{trainMeanSd} \tab \tab mean and standard deviation of training data. \cr
  \code{Cy} \tab \tab sample covariance for response variables. \cr
  \code{E} \tab \tab error matrix. \cr
  \code{Ie} \tab \tab identity matrix for setting E for all observations. \cr
  \code{K} \tab \tab kernel matrix. \cr
  \code{kernel} \tab \tab a name of the kernel function. \cr
  \code{sigma} \tab \tab value of sigma. \cr
  \code{corlen} \tab \tab value of correlation length. \cr
  \code{errorvar} \tab \tab errorvar noise variance (a.k.a. nugget parameter). \cr
  \code{meanf} \tab \tab name of the mean function \cr
  \code{XColName} \tab \tab column names of predictor variables. \cr
  \code{YColName} \tab \tab column names of response variable(s). \cr
  \code{XRowName} \tab \tab row names of predictor variables. \cr
}
}
\description{
\code{mgpr} is used to fit a Multivariate Gaussian Process Regression model.
}
\details{
What to put here? Do we need this?
}
\examples{
data(mgprdata)
m <- mgpr(datay = mgprdata[, 1:5], datax = mgprdata[, 6:29],
kernel = "matern32", kernpar = list(sigma = 1, corlen = 5, errorvar = 0.1))
}
\references{
Varvia, P., Lahivaara, T., Maltamo, M., Packalen, P. and Seppanen, A. 2019.
Gaussian process regression for forest attribute estimation from airborne laser scanning
data. IEEE Transactions on Geoscience and Remote Sensing 57(6): 3361-3369.

R package article to appear.
}
