% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Summarizing an mgpr model}
\usage{
summary(mgpr)
}
\arguments{
\item{mgpr}{a \code{mgpr} model object, usually, 
a result of a call to \code{\link{mgpr}}.}
}
\value{
A summary of the mgpr model will be printed to the R console.
}
\description{
\code{summary} method for class "\code{mgpr}".
}
\examples{
data(mgprdata)
m <- mgpr(
  datay = mgprdata[, 1:3], datax = mgprdata[, 4:39],
  kernel = "matern32", kernpar = list(sigma = 1, corlen = 5, errorvar = 0.1)
)
summary(m)
}
